instrumentations/bedrock/pyproject.toml
[project]
name = "llm-tracekit-bedrock"
version = "1.0.0"
description = "OpenTelemetry instrumentation for AWS Bedrock."
readme = "../../README.md"
authors = [
    { name="Coralogix Ltd.", email="info@coralogix.com" },
]
license = "Apache-2.0"
license-files = ["../../LICENSE"]
requires-python = ">=3.10,<3.14"
dependencies = [
    "llm-tracekit-core>=1.0.0",
    "opentelemetry-instrumentation>=0.53b1",
    "boto3>=1.36.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: OpenTelemetry :: Instrumentations",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://coralogix.com"
repository = "https://github.com/coralogix/llm-tracekit.git"

[project.entry-points.opentelemetry_instrumentor]
boto3 = "llm_tracekit.bedrock:BedrockInstrumentor"

[dependency-groups]
dev = [
    "assertpy>=1.1",
    "botocore-stubs>=1.38.4",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-vcr>=1.0.2",
    "pyyaml>=6.0.2",
    "ruff>=0.9.7",
    "types-boto3>=1.38.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_tracekit"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src/llm_tracekit",
]

[tool.uv.sources]
llm-tracekit-core = { workspace = true }

